ai/requirements.txt
fastapi==0.115.11
python-multipart==0.0.6
uvicorn==0.34.0
celery==5.3.6
redis
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
minio==7.2.4
aio-pika>=9.0.0,<10.0.0

librosa==0.9.2
numpy==2.1.3
scipy==1.15.2
fastdtw==0.3.4
soundfile==0.13.1
